pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mgen3d"
version = "0.0.1"
dependencies = [
  "absl-py >= 2.0.0",
  "coverage",
  "opencv-python >= 4.8.0.0",
  "nerfacc >= 0.5.3",
  "ninja",
  "numpy  >= 1.19.5",
  "scipy >= 1.10.0",
  "pillow >= 9.0.0",
  "tqdm > 4.60.0",
  "torch >= 1.12.0",
  "trimesh >= 4.0.5",
  "pytest >= 7.0.0",
  "pyglet < 2.0.0",
  "tinycudann @ git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch",
  "clip @ git+https://github.com/openai/CLIP.git"
]
requires-python = ">=3.8"
authors = [
  {name = "Bryson Jones", email = "bkjones97@gmail.com"},
  {name = "Guglielmo Daddi", email = "TODO@gmail.com"}
]
maintainers = [
  {name = "Bryson Jones", email = "bkjones97@gmail.com"},
  {name = "Guglielmo Daddi", email = "TODO@gmail.com"}
]
description = "A python library for generating 3D Mesh Models with text and image guidance"
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["generative", "model", "mesh", "diffusion", "LLM", "VLM", "NeRF", "3D"]

[project.optional-dependencies]
# TODO: Add optional dependencies here

[project.urls]
# TODO:
# Homepage = "https://example.com"
# Documentation = "https://readthedocs.org"
# Repository = "https://github.com/me/spam.git"
# "Bug Tracker" = "https://github.com/me/spam/issues"
# Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[project.scripts]
mgen3d-viz-test = "mgen3d.pose_generator_utils:main"
train_full_nerf = "mgen3d.pipelines.nerf_pipeline:main"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
